+++ desktop/geogebra/awt/GeneralPath.java
@@ -119,7 +119,7 @@ public class GeneralPath extends geogebra.common.awt.GeneralPath implements geog
 		return impl.contains(p.getX(), p.getY());
 	}
 	@Override
-	public void curveTo(double f, double g, double h, double i, double j, double k) {
+	public void curveTo(float f, float g, float h, float i, float j, float k) {
 		impl.curveTo(f, g, h, i, j, k);
 		
 	}
